refactor: split deploy preview ci helpers
