+++ genie/ci-workflow.ts
@@ -31,6 +31,15 @@ import {
   type GitHubWorkflowArgs,
 } from '../packages/@overeng/genie/src/runtime/mod.ts'
 import { RUNNER_PROFILES, type RunnerProfile } from './ci.ts'
+import {
+  netlifyDeployStep as buildNetlifyDeployStep,
+  netlifyStorybookCommentStep as buildNetlifyStorybookCommentStep,
+} from './deploy-preview/netlify.ts'
+import {
+  type VercelProject,
+  vercelDeployJobs as buildVercelDeployJobs,
+  vercelDeployStep as buildVercelDeployStep,
+} from './deploy-preview/vercel.ts'
 
 export { RUNNER_PROFILES, type RunnerProfile }
 
@@ -840,330 +849,6 @@ export const deployCommentStep = (opts: {
   ].join('\n'),
 })
 
-const deployTargetEnvSuffix = (name: string) =>
-  name
-    .toUpperCase()
-    .replaceAll('-', '_')
-    .replaceAll(/[^A-Z0-9_]/g, '')
-
-const deployPreviewManagedMarker = '<!-- deploy-preview-comment:managed -->'
-const deployPreviewStatePrefix = '<!-- deploy-preview-comment:state\n'
-const deployPreviewStateSuffix = '\n-->'
-
-const vercelDeployCommentStep = (opts: {
-  commentTitle: string
-  noRowsMessage: string
-  projects: readonly Pick<VercelProject, 'name' | 'label'>[]
-}) => {
-  const projects = opts.projects.map((project) => ({
-    name: project.name,
-    displayName: project.label ?? project.name,
-    envSuffix: deployTargetEnvSuffix(project.name),
-  }))
-
-  const renderCommentScript = [
-    "import { readFileSync, writeFileSync } from 'node:fs'",
-    '',
-    `const commentTitle = ${JSON.stringify(opts.commentTitle)}`,
-    `const noRowsMessage = ${JSON.stringify(opts.noRowsMessage)}`,
-    `const projects = ${JSON.stringify(projects)}`,
-    `const managedMarker = ${JSON.stringify(deployPreviewManagedMarker)}`,
-    `const statePrefix = ${JSON.stringify(deployPreviewStatePrefix)}`,
-    `const stateSuffix = ${JSON.stringify(deployPreviewStateSuffix)}`,
-    `const stateTag = 'deploy-preview-comment-state'`,
-    `const schemaVersion = 1`,
-    `const timeZone = 'Europe/Berlin'`,
-    `const maxCommits = 50`,
-    '',
-    'const [commentsPath, commentBodyPath, summaryPath, commentIdPath] = process.argv.slice(2)',
-    '',
-    'const fail = (message) => {',
-    '  throw new Error(message)',
-    '}',
-    '',
-    'const expectObject = (value, path) => {',
-    "  if (typeof value !== 'object' || value === null || Array.isArray(value)) {",
-    '    fail(`${path} must be an object`)',
-    '  }',
-    '  return value',
-    '}',
-    '',
-    'const expectArray = (value, path) => {',
-    '  if (!Array.isArray(value)) {',
-    '    fail(`${path} must be an array`)',
-    '  }',
-    '  return value',
-    '}',
-    '',
-    'const expectString = (value, path) => {',
-    "  if (typeof value !== 'string' || value.length === 0) {",
-    '    fail(`${path} must be a non-empty string`)',
-    '  }',
-    '  return value',
-    '}',
-    '',
-    'const expectExactKeys = (record, keys, path) => {',
-    '  const actualKeys = Object.keys(record).sort()',
-    '  const expectedKeys = [...keys].sort()',
-    '  if (actualKeys.length !== expectedKeys.length || actualKeys.some((key, index) => key !== expectedKeys[index])) {',
-    '    fail(`${path} keys must be exactly: ${expectedKeys.join(", ")}`)',
-    '  }',
-    '}',
-    '',
-    'const expectUrl = (value, path) => {',
-    '  const string = expectString(value, path)',
-    "  if (!string.startsWith('https://')) {",
-    '    fail(`${path} must start with https://`)',
-    '  }',
-    '  return string',
-    '}',
-    '',
-    'const expectIsoUtc = (value, path) => {',
-    '  const string = expectString(value, path)',
-    '  if (!/^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$/.test(string) || Number.isNaN(Date.parse(string))) {',
-    '    fail(`${path} must be a valid UTC ISO timestamp`)',
-    '  }',
-    '  return string',
-    '}',
-    '',
-    'const validateTarget = (value, path) => {',
-    '  const record = expectObject(value, path)',
-    "  expectExactKeys(record, ['target', 'displayName', 'finalUrl', 'rawDeployUrl', 'deployedAtUtc'], path)",
-    '  return {',
-    '    target: expectString(record.target, `${path}.target`),',
-    '    displayName: expectString(record.displayName, `${path}.displayName`),',
-    '    finalUrl: expectUrl(record.finalUrl, `${path}.finalUrl`),',
-    '    rawDeployUrl: expectUrl(record.rawDeployUrl, `${path}.rawDeployUrl`),',
-    '    deployedAtUtc: expectIsoUtc(record.deployedAtUtc, `${path}.deployedAtUtc`),',
-    '  }',
-    '}',
-    '',
-    'const validateCommit = (value, path) => {',
-    '  const record = expectObject(value, path)',
-    "  expectExactKeys(record, ['commitSha', 'modeLabel', 'targets'], path)",
-    '  const targets = expectArray(record.targets, `${path}.targets`).map((target, index) =>',
-    '    validateTarget(target, `${path}.targets[${index}]`),',
-    '  )',
-    '  return {',
-    '    commitSha: expectString(record.commitSha, `${path}.commitSha`),',
-    '    modeLabel: expectString(record.modeLabel, `${path}.modeLabel`),',
-    '    targets,',
-    '  }',
-    '}',
-    '',
-    'const validateState = (value) => {',
-    "  const record = expectObject(value, 'state')",
-    "  expectExactKeys(record, ['_tag', 'schemaVersion', 'timeZone', 'targetOrder', 'commits'], 'state')",
-    '  if (record._tag !== stateTag) fail(`state._tag must be ${stateTag}`)',
-    '  if (record.schemaVersion !== schemaVersion) fail(`state.schemaVersion must be ${schemaVersion}`)',
-    '  if (record.timeZone !== timeZone) fail(`state.timeZone must be ${timeZone}`)',
-    '  const targetOrder = expectArray(record.targetOrder, "state.targetOrder").map((target, index) =>',
-    '    expectString(target, `state.targetOrder[${index}]`),',
-    '  )',
-    '  const commits = expectArray(record.commits, "state.commits").map((commit, index) =>',
-    '    validateCommit(commit, `state.commits[${index}]`),',
-    '  )',
-    '  return { _tag: stateTag, schemaVersion, timeZone, targetOrder, commits }',
-    '}',
-    '',
-    'const formatBerlin = (isoUtc) => {',
-    "  const formatter = new Intl.DateTimeFormat('en-GB', {",
-    '    timeZone,',
-    "    year: 'numeric',",
-    "    month: '2-digit',",
-    "    day: '2-digit',",
-    "    hour: '2-digit',",
-    "    minute: '2-digit',",
-    '    hour12: false,',
-    "    timeZoneName: 'short',",
-    '  })',
-    '  const parts = Object.fromEntries(',
-    '    formatter.formatToParts(new Date(isoUtc)).filter((part) => part.type !== "literal").map((part) => [part.type, part.value]),',
-    '  )',
-    '  return `${parts.year}-${parts.month}-${parts.day} ${parts.hour}:${parts.minute} ${parts.timeZoneName}`',
-    '}',
-    '',
-    'const extractState = (body) => {',
-    '  const start = body.indexOf(statePrefix)',
-    '  if (start === -1) return undefined',
-    '  const end = body.indexOf(stateSuffix, start + statePrefix.length)',
-    '  if (end === -1) fail("existing managed comment is missing the state suffix marker")',
-    '  const raw = body.slice(start + statePrefix.length, end)',
-    '  return validateState(JSON.parse(raw))',
-    '}',
-    '',
-    'const renderHistoryUrl = (target) =>',
-    '  target.finalUrl === target.rawDeployUrl',
-    '    ? target.rawDeployUrl',
-    '    : `${target.rawDeployUrl}<br><sub>Alias: ${target.finalUrl}</sub>`',
-    '',
-    'const currentTargets = projects.flatMap((project) => {',
-    '  const finalUrl = process.env[`DEPLOY_FINAL_URL_${project.envSuffix}`] ?? ""',
-    '  const rawDeployUrl = process.env[`DEPLOY_RAW_DEPLOY_URL_${project.envSuffix}`] ?? finalUrl',
-    '  const deployedAtUtc = process.env[`DEPLOYED_AT_UTC_${project.envSuffix}`] ?? ""',
-    '  if (finalUrl.length === 0 || deployedAtUtc.length === 0) return []',
-    '  return [{',
-    '    target: project.name,',
-    '    displayName: project.displayName,',
-    '    finalUrl,',
-    '    rawDeployUrl: rawDeployUrl.length === 0 ? finalUrl : rawDeployUrl,',
-    '    deployedAtUtc,',
-    '  }]',
-    '})',
-    '',
-    'const dedupedCurrentTargets = [...new Map(currentTargets.map((target) => [target.target, target])).values()]',
-    'dedupedCurrentTargets.forEach((target, index) => validateTarget(target, `currentTargets[${index}]`))',
-    '',
-    'if (dedupedCurrentTargets.length === 0) {',
-    '  writeFileSync(summaryPath, `## ${commentTitle}\\n\\n${noRowsMessage}\\n`)',
-    "  writeFileSync(commentBodyPath, '')",
-    "  writeFileSync(commentIdPath, '')",
-    '  process.exit(0)',
-    '}',
-    '',
-    "const comments = JSON.parse(readFileSync(commentsPath, 'utf8'))",
-    'if (!Array.isArray(comments)) fail("comments response must be an array")',
-    '',
-    'const existingComment = comments',
-    '  .filter(',
-    '    (comment) =>',
-    '      typeof comment?.body === "string" &&',
-    '      (comment.body.includes(managedMarker) || comment.body.startsWith(`## ${commentTitle}`)),',
-    '  )',
-    '  .at(-1)',
-    '',
-    'const existingState = existingComment ? extractState(existingComment.body) : undefined',
-    'const priorState = existingState ?? { _tag: stateTag, schemaVersion, timeZone, targetOrder: [], commits: [] }',
-    '',
-    'const commitSha = expectString(process.env.DEPLOY_COMMIT_SHA, "DEPLOY_COMMIT_SHA")',
-    'const modeLabel = expectString(process.env.DEPLOY_LABEL, "DEPLOY_LABEL")',
-    '',
-    'const nextTargetOrder = [...new Set([...dedupedCurrentTargets.map((target) => target.target), ...priorState.targetOrder])]',
-    'const nextState = {',
-    '  _tag: stateTag,',
-    '  schemaVersion,',
-    '  timeZone,',
-    '  targetOrder: nextTargetOrder,',
-    '  commits: [',
-    '    { commitSha, modeLabel, targets: dedupedCurrentTargets },',
-    '    ...priorState.commits.filter((commit) => commit.commitSha !== commitSha),',
-    '  ].slice(0, maxCommits),',
-    '}',
-    '',
-    'const latestByTarget = new Map()',
-    'for (const commit of nextState.commits) {',
-    '  for (const target of commit.targets) {',
-    '    if (!latestByTarget.has(target.target)) {',
-    '      latestByTarget.set(target.target, target)',
-    '    }',
-    '  }',
-    '}',
-    '',
-    'const renderCommitTimestamp = (commit) => {',
-    '  const latestIso = commit.targets.reduce((currentLatest, target) =>',
-    '    Date.parse(target.deployedAtUtc) > Date.parse(currentLatest) ? target.deployedAtUtc : currentLatest,',
-    '  commit.targets[0]?.deployedAtUtc ?? new Date(0).toISOString())',
-    '  return formatBerlin(latestIso)',
-    '}',
-    '',
-    'const visibleLines = [',
-    '  `## ${commentTitle}`,',
-    "  '',",
-    '  `| Target | Latest URL | Last Deploy (Europe/Berlin) |`,',
-    '  `| --- | --- | --- |`,',
-    '  ...nextState.targetOrder.flatMap((targetName) => {',
-    '    const target = latestByTarget.get(targetName)',
-    '    return target === undefined',
-    '      ? []',
-    '      : [`| ${target.displayName} | ${target.finalUrl} | ${formatBerlin(target.deployedAtUtc)} |`]',
-    '  }),',
-    ']',
-    '',
-    'if (nextState.commits.length > 0) {',
-    "  visibleLines.push('', '<details>', '<summary>Per-Commit Deploy History</summary>', '')",
-    '  for (const commit of nextState.commits) {',
-    '    visibleLines.push(`### Commit \\`${commit.commitSha.slice(0, 7)}\\` · ${renderCommitTimestamp(commit)}`)',
-    "    visibleLines.push('')",
-    '    visibleLines.push(`| Target | URL |`) ',
-    '    visibleLines.push(`| --- | --- |`)',
-    '    for (const targetName of nextState.targetOrder) {',
-    '      const target = commit.targets.find((entry) => entry.target === targetName)',
-    '      if (target !== undefined) {',
-    '        visibleLines.push(`| ${target.displayName} | ${renderHistoryUrl(target)} |`)',
-    '      }',
-    '    }',
-    "    visibleLines.push('')",
-    '  }',
-    "  visibleLines.push('</details>')",
-    '}',
-    '',
-    'const visibleBody = `${visibleLines.join("\\n")}\\n`',
-    'const hiddenState = `${managedMarker}\\n${statePrefix}${JSON.stringify(nextState, undefined, 2)}${stateSuffix}`',
-    'const fullBody = `${visibleBody}\\n${hiddenState}\\n`',
-    '',
-    'writeFileSync(summaryPath, visibleBody)',
-    'writeFileSync(commentBodyPath, fullBody)',
-    "writeFileSync(commentIdPath, existingComment ? String(existingComment.id) : '')",
-    '',
-  ].join('\n')
-
-  return {
-    name: 'Post deploy URLs',
-    if: 'always() && !cancelled()',
-    shell: 'bash' as const,
-    env: {
-      GH_TOKEN: '${{ github.token }}',
-      GH_REPO: '${{ github.repository }}',
-      ...Object.fromEntries(
-        projects.flatMap((project) => [
-          [
-            `DEPLOY_FINAL_URL_${project.envSuffix}`,
-            `\${{ needs.deploy-${project.name}.outputs.final_url }}`,
-          ],
-          [
-            `DEPLOY_RAW_DEPLOY_URL_${project.envSuffix}`,
-            `\${{ needs.deploy-${project.name}.outputs.raw_deploy_url }}`,
-          ],
-          [
-            `DEPLOYED_AT_UTC_${project.envSuffix}`,
-            `\${{ needs.deploy-${project.name}.outputs.deployed_at_utc }}`,
-          ],
-        ]),
-      ),
-    },
-    run: [
-      deployModeScript,
-      'if [ "${{ github.event_name }}" = "pull_request" ]; then',
-      '  commit_sha="${{ github.event.pull_request.head.sha }}"',
-      'else',
-      '  commit_sha="${{ github.sha }}"',
-      'fi',
-      'export DEPLOY_LABEL="$label"',
-      'export DEPLOY_COMMIT_SHA="$commit_sha"',
-      'comments_json="/tmp/deploy-comments.json"',
-      'if [ "${{ github.event_name }}" = "pull_request" ]; then',
-      '  export NIX_CONFIG="${NIX_CONFIG:+$NIX_CONFIG$\'\\n\'}access-tokens = github.com=${GH_TOKEN}"',
-      '  nix run nixpkgs#gh -- api "repos/$GH_REPO/issues/${{ github.event.pull_request.number }}/comments" --paginate > "$comments_json"',
-      'else',
-      '  printf \'[]\' > "$comments_json"',
-      'fi',
-      "cat > /tmp/render-deploy-comment.mjs <<'EOF'",
-      renderCommentScript,
-      'EOF',
-      'nix run nixpkgs#nodejs_24 -- /tmp/render-deploy-comment.mjs "$comments_json" /tmp/comment.md /tmp/summary.md /tmp/comment-id.txt',
-      'cat /tmp/summary.md >> "$GITHUB_STEP_SUMMARY"',
-      'if [ "${{ github.event_name }}" = "pull_request" ]; then',
-      '  comment_id="$(cat /tmp/comment-id.txt)"',
-      '  if [ -n "$comment_id" ]; then',
-      '    nix run nixpkgs#gh -- api "repos/$GH_REPO/issues/comments/$comment_id" --method PATCH --field body="$(cat /tmp/comment.md)" > /dev/null',
-      '  else',
-      '    nix run nixpkgs#gh -- api "repos/$GH_REPO/issues/${{ github.event.pull_request.number }}/comments" --method POST --field body="$(cat /tmp/comment.md)" > /dev/null',
-      '  fi',
-      'fi',
-    ].join('\n'),
-  }
-}
-
 /**
  * Step that dispatches `upstream-changed` repository_dispatch to a target repo.
  * Add this to upstream CI workflows so merges to main trigger downstream alignment.
@@ -1213,66 +898,8 @@ export const notifyAlignmentJob = (opts: {
  * Captures final/raw deploy URLs plus deploy completion time and exports them
  * to both GITHUB_ENV and GITHUB_OUTPUT.
  */
-export const vercelDeployStep = (project: { name: string; urlEnvKey?: string }) => {
-  const urlEnvKey =
-    project.urlEnvKey ??
-    `VERCEL_DEPLOY_URL_${project.name
-      .toUpperCase()
-      .replaceAll('-', '_')
-      .replaceAll(/[^A-Z0-9_]/g, '')}`
-
-  return {
-    id: 'deploy',
-    name: `Deploy ${project.name} to Vercel`,
-    shell: 'bash' as const,
-    run: [
-      'if [ -z "${VERCEL_TOKEN:-}" ]; then',
-      '  echo "::error::VERCEL_TOKEN is not set"',
-      '  exit 1',
-      'fi',
-      'tmp_log="$(mktemp)"',
-      'if [ "${{ github.event_name }}" = "pull_request" ]; then',
-      `  ${runDevenvTasksBefore(`vercel:deploy:${project.name}`, '--show-output', '--input', 'type=pr', '--input', 'pr=${{ github.event.pull_request.number }}')} 2>&1 | tee "$tmp_log"`,
-      'else',
-      `  ${runDevenvTasksBefore(`vercel:deploy:${project.name}`, '--show-output', '--input', 'type=prod')} 2>&1 | tee "$tmp_log"`,
-      'fi',
-      'deploy_exit=${PIPESTATUS[0]}',
-      'if [ "$deploy_exit" -ne 0 ]; then exit "$deploy_exit"; fi',
-      `metadata_json=$(grep -F 'DEPLOY_TASK_METADATA: ' "$tmp_log" | sed 's/^.*DEPLOY_TASK_METADATA: //' | tail -n 1 || true)`,
-      'if [ -n "$metadata_json" ]; then',
-      '  final_url=$(printf "%s" "$metadata_json" | jq -r \'.finalUrl // empty\')',
-      '  raw_deploy_url=$(printf "%s" "$metadata_json" | jq -r \'.rawDeployUrl // empty\')',
-      '  deployed_at_utc=$(printf "%s" "$metadata_json" | jq -r \'.deployedAtUtc // empty\')',
-      'else',
-      '  final_url=""',
-      '  raw_deploy_url=""',
-      '  deployed_at_utc=""',
-      'fi',
-      'if [ -z "$final_url" ]; then',
-      `  final_url=$(grep -Eo 'Vercel deploy URL: https://[^[:space:]"]+' "$tmp_log" | sed 's/^Vercel deploy URL: //' | tail -n 1 || true)`,
-      'fi',
-      'if [ -z "$final_url" ]; then',
-      `  final_url=$(grep -oE 'https://[^[:space:]"]+' "$tmp_log" | grep -E 'vercel\\.(app|com)' | tail -n 1 || true)`,
-      'fi',
-      'if [ -z "$raw_deploy_url" ]; then',
-      '  raw_deploy_url="$final_url"',
-      'fi',
-      'if [ -z "$deployed_at_utc" ]; then',
-      '  deployed_at_utc="$(date -u +%Y-%m-%dT%H:%M:%SZ)"',
-      'fi',
-      'if [ -n "$final_url" ]; then',
-      `  echo "${urlEnvKey}=$final_url" >> "$GITHUB_ENV"`,
-      '  echo "final_url=$final_url" >> "$GITHUB_OUTPUT"',
-      '  echo "deploy_url=$final_url" >> "$GITHUB_OUTPUT"',
-      'fi',
-      'if [ -n "$raw_deploy_url" ]; then',
-      '  echo "raw_deploy_url=$raw_deploy_url" >> "$GITHUB_OUTPUT"',
-      'fi',
-      'echo "deployed_at_utc=$deployed_at_utc" >> "$GITHUB_OUTPUT"',
-      'rm -f "$tmp_log"',
-    ].join('\n'),
-  }
-}
+export const vercelDeployStep = (project: { name: string; urlEnvKey?: string }) =>
+  buildVercelDeployStep(project, runDevenvTasksBefore)
 
 /**
  * Configure git author so Vercel Deployment Protection
@@ -1287,14 +914,6 @@ export const vercelGitAuthorStep = (opts: { name: string; email: string }) => ({
     'git commit --amend --no-edit --reset-author',
   ].join('\n'),
 })
-
-type VercelProject = {
-  name: string
-  urlEnvKey?: string
-  projectIdEnv: string
-  label?: string
-  stepsBeforeDeploy?: readonly Record<string, unknown>[]
-}
 
 /**
  * Generate Vercel deploy jobs and optionally a combined comment collector job.
@@ -1324,67 +943,14 @@ export const vercelDeployJobs = (opts: {
   commentTitle?: string
   noRowsMessage?: string
 }): Record<string, Record<string, unknown>> => {
-  const deployCondition =
-    opts.deployCondition ??
-    [
-      'always()',
-      `(github.event_name == 'schedule' || (${(opts.needs ?? []).map((j) => `needs.${j}.result == 'success'`).join(' && ')}))`,
-    ].join(' && ')
-
-  const deployJobNames = opts.projects.map((p) => `deploy-${p.name}`)
-
-  const deployJobs = Object.fromEntries(
-    opts.projects.map((project) => [
-      `deploy-${project.name}`,
-      {
-        ...(opts.needs !== undefined && opts.needs.length > 0 ? { needs: [...opts.needs] } : {}),
-        if: deployCondition,
-        'runs-on': [...opts.runner],
-        defaults: bashShellDefaults,
-        outputs: {
-          final_url: '${{ steps.deploy.outputs.final_url }}',
-          raw_deploy_url: '${{ steps.deploy.outputs.raw_deploy_url }}',
-          deployed_at_utc: '${{ steps.deploy.outputs.deployed_at_utc }}',
-          deploy_url: '${{ steps.deploy.outputs.deploy_url }}',
-        },
-        env: {
-          ...opts.env,
-          [project.projectIdEnv]:
-            opts.env[project.projectIdEnv] ?? `\${{ secrets.${project.projectIdEnv} }}`,
-        },
-        steps: [
-          ...opts.baseSteps,
-          ...(project.stepsBeforeDeploy ?? []),
-          vercelDeployStep(project),
-          ...(opts.extraSteps ?? []),
-        ],
-      },
-    ]),
-  )
-
-  if (opts.includeComment === false) {
-    return deployJobs
-  }
-
-  /** Collect URLs from parallel deploy jobs and post one combined comment. */
-  const commentJob = {
-    needs: deployJobNames,
-    if: 'always() && !cancelled()',
-    permissions: deployCommentPermissions,
-    'runs-on': linuxX64Runner,
-    steps: [
-      vercelDeployCommentStep({
-        commentTitle: opts.commentTitle ?? 'Deploy Preview',
-        projects: opts.projects,
-        noRowsMessage: opts.noRowsMessage ?? 'No deploy URLs detected.',
-      }),
-    ],
-  }
-
-  return {
-    ...deployJobs,
-    'post-deploy-comment': commentJob,
-  }
+  return buildVercelDeployJobs({
+    ...opts,
+    runDevenvTasksBefore,
+    deployModeScript,
+    deployCommentPermissions,
+    bashShellDefaults,
+    commentRunner: linuxX64Runner,
+  })
 }
 
 // =============================================================================
@@ -1396,55 +962,7 @@ export const vercelDeployJobs = (opts: {
  * Runs `netlify:deploy` with prod/PR mode based on the event trigger.
  * Gracefully skips if NETLIFY_AUTH_TOKEN is not available.
  */
-export const netlifyDeployStep = () => ({
-  id: 'deploy',
-  name: 'Deploy storybooks to Netlify',
-  shell: 'bash' as const,
-  run: [
-    'tmp_log="/tmp/netlify-storybook-deploy.log"',
-    'rm -f "$tmp_log"',
-    'if [ -z "${NETLIFY_AUTH_TOKEN:-}" ]; then',
-    '  echo "::notice::Skipping Netlify deploy (NETLIFY_AUTH_TOKEN not available)"',
-    '  exit 0',
-    'fi',
-    'if [ "${{ github.event_name }}" = "push" ] && [ "${{ github.ref }}" = "refs/heads/main" ]; then',
-    `  ${runDevenvTasksBefore('netlify:deploy', '--show-output', '--input', 'type=prod')} 2>&1 | tee "$tmp_log"`,
-    'elif [ "${{ github.event_name }}" = "pull_request" ]; then',
-    `  ${runDevenvTasksBefore('netlify:deploy', '--show-output', '--input', 'type=pr', '--input', 'pr=${{ github.event.pull_request.number }}')} 2>&1 | tee "$tmp_log"`,
-    'fi',
-    'deploy_exit=${PIPESTATUS[0]}',
-    'if [ "$deploy_exit" -ne 0 ]; then exit "$deploy_exit"; fi',
-    "cat > /tmp/extract-netlify-preview-packages.mjs <<'EOF'",
-    [
-      "import { readFileSync } from 'node:fs'",
-      '',
-      'const [logPath] = process.argv.slice(2)',
-      'const prefixes = ["DEPLOY_TASK_METADATA: ", "NETLIFY_DEPLOY_METADATA: "]',
-      'const packages = readFileSync(logPath, "utf8")',
-      '  .split("\\n")',
-      '  .flatMap((line) => {',
-      '    const prefix = prefixes.find((candidate) => line.includes(candidate))',
-      '    if (!prefix) return []',
-      '    const rawRecord = JSON.parse(line.slice(line.indexOf(prefix) + prefix.length))',
-      '    if (rawRecord.provider === "netlify") {',
-      '      return [{',
-      '        packageName: rawRecord.displayName ?? rawRecord.target,',
-      '        rawDeployUrl: rawRecord.rawDeployUrl,',
-      '        finalUrl: rawRecord.finalUrl,',
-      '        deployedAtUtc: rawRecord.deployedAtUtc,',
-      '      }]',
-      '    }',
-      '    if ("packageName" in rawRecord) return [rawRecord]',
-      '    return []',
-      '  })',
-      'const dedupedPackages = [...new Map(packages.map((pkg) => [pkg.packageName, pkg])).values()]',
-      'process.stdout.write(JSON.stringify(dedupedPackages))',
-    ].join('\n'),
-    'EOF',
-    'packages_json="$(node /tmp/extract-netlify-preview-packages.mjs "$tmp_log")"',
-    'printf \'packages_json<<__NETLIFY_PACKAGES__\\n%s\\n__NETLIFY_PACKAGES__\\n\' "$packages_json" >> "$GITHUB_OUTPUT"',
-  ].join('\n'),
-})
+export const netlifyDeployStep = () => buildNetlifyDeployStep(runDevenvTasksBefore)
 
 /**
  * Combined deploy comment step for Netlify storybook previews.
@@ -1453,318 +971,5 @@ export const netlifyDeployStep = () => ({
  *
  * @param site - Netlify site name (e.g. 'overeng-utils')
  */
-export const netlifyStorybookCommentStep = (site: string) => ({
-  name: 'Post storybook preview comment',
-  if: '${{ !cancelled() }}',
-  env: {
-    GH_TOKEN: '${{ github.token }}',
-    GH_REPO: '${{ github.repository }}',
-    STORYBOOK_PREVIEW_PACKAGES_JSON: '${{ steps.deploy.outputs.packages_json }}',
-  },
-  run: [
-    `site="${site}"`,
-    deployModeScript,
-    '# Set Netlify branch-deploy suffix based on mode',
-    'if [ "$label" = "prod" ]; then suffix=""; else suffix="-pr-${{ github.event.pull_request.number }}"; fi',
-    'if [ "${{ github.event_name }}" = "pull_request" ]; then',
-    '  commit_sha="${{ github.event.pull_request.head.sha }}"',
-    'else',
-    '  commit_sha="${{ github.sha }}"',
-    'fi',
-    'export STORYBOOK_PREVIEW_LABEL="$label"',
-    'export STORYBOOK_PREVIEW_COMMIT_SHA="$commit_sha"',
-    'export STORYBOOK_PREVIEW_SITE="$site"',
-    'export STORYBOOK_PREVIEW_SUFFIX="$suffix"',
-    'export STORYBOOK_PREVIEW_DEPLOYED_AT_UTC="$(date -u +%Y-%m-%dT%H:%M:%SZ)"',
-    'comments_json="/tmp/storybook-preview-comments.json"',
-    'if [ "${{ github.event_name }}" = "pull_request" ]; then',
-    `  export NIX_CONFIG="\${NIX_CONFIG:+$NIX_CONFIG$'\\n'}access-tokens = github.com=\${GH_TOKEN}"`,
-    '  nix run nixpkgs#gh -- api "repos/$GH_REPO/issues/${{ github.event.pull_request.number }}/comments" --paginate > "$comments_json"',
-    'else',
-    '  printf \'[]\' > "$comments_json"',
-    'fi',
-    "cat > /tmp/render-storybook-preview-comment.mjs <<'EOF'",
-    [
-      "import { existsSync, readdirSync, readFileSync, writeFileSync } from 'node:fs'",
-      '',
-      'const commentTitle = "Storybook Previews"',
-      'const noRowsMessage = "No storybooks were deployed."',
-      'const managedMarker = "<!-- deploy-preview-comment:managed -->"',
-      'const statePrefix = "<!-- deploy-preview-comment:state\\n"',
-      'const stateSuffix = "\\n-->"',
-      "const stateTag = 'storybook-preview-comment-state'",
-      'const schemaVersion = 3',
-      "const timeZone = 'Europe/Berlin'",
-      'const maxCommits = 50',
-      '',
-      'const [commentsPath, commentBodyPath, summaryPath, commentIdPath] = process.argv.slice(2)',
-      '',
-      'const fail = (message) => {',
-      '  throw new Error(message)',
-      '}',
-      '',
-      'const expectObject = (value, path) => {',
-      "  if (typeof value !== 'object' || value === null || Array.isArray(value)) {",
-      '    fail(`${path} must be an object`)',
-      '  }',
-      '  return value',
-      '}',
-      '',
-      'const expectArray = (value, path) => {',
-      '  if (!Array.isArray(value)) {',
-      '    fail(`${path} must be an array`)',
-      '  }',
-      '  return value',
-      '}',
-      '',
-      'const expectString = (value, path) => {',
-      "  if (typeof value !== 'string' || value.length === 0) {",
-      '    fail(`${path} must be a non-empty string`)',
-      '  }',
-      '  return value',
-      '}',
-      '',
-      'const expectExactKeys = (record, keys, path) => {',
-      '  const actualKeys = Object.keys(record).sort()',
-      '  const expectedKeys = [...keys].sort()',
-      '  if (actualKeys.length !== expectedKeys.length || actualKeys.some((key, index) => key !== expectedKeys[index])) {',
-      '    fail(`${path} keys must be exactly: ${expectedKeys.join(", ")}`)',
-      '  }',
-      '}',
-      '',
-      'const expectUrl = (value, path) => {',
-      '  const string = expectString(value, path)',
-      "  if (!string.startsWith('https://')) {",
-      '    fail(`${path} must start with https://`)',
-      '  }',
-      '  return string',
-      '}',
-      '',
-      'const expectIsoUtc = (value, path) => {',
-      '  const string = expectString(value, path)',
-      '  if (!/^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$/.test(string) || Number.isNaN(Date.parse(string))) {',
-      '    fail(`${path} must be a valid UTC ISO timestamp`)',
-      '  }',
-      '  return string',
-      '}',
-      '',
-      'const validatePackageV1 = (value, path) => {',
-      '  const record = expectObject(value, path)',
-      "  expectExactKeys(record, ['packageName', 'url'], path)",
-      '  return {',
-      '    packageName: expectString(record.packageName, `${path}.packageName`),',
-      '    rawDeployUrl: expectUrl(record.url, `${path}.url`),',
-      '    finalUrl: expectUrl(record.url, `${path}.url`),',
-      '    deployedAtUtc: undefined,',
-      '  }',
-      '}',
-      '',
-      'const validatePackageV2 = (value, path) => {',
-      '  const record = expectObject(value, path)',
-      "  expectExactKeys(record, ['packageName', 'rawDeployUrl', 'finalUrl'], path)",
-      '  return {',
-      '    packageName: expectString(record.packageName, `${path}.packageName`),',
-      '    rawDeployUrl: expectUrl(record.rawDeployUrl, `${path}.rawDeployUrl`),',
-      '    finalUrl: expectUrl(record.finalUrl, `${path}.finalUrl`),',
-      '    deployedAtUtc: undefined,',
-      '  }',
-      '}',
-      '',
-      'const validatePackageV3 = (value, path) => {',
-      '  const record = expectObject(value, path)',
-      "  expectExactKeys(record, ['packageName', 'rawDeployUrl', 'finalUrl', 'deployedAtUtc'], path)",
-      '  return {',
-      '    packageName: expectString(record.packageName, `${path}.packageName`),',
-      '    rawDeployUrl: expectUrl(record.rawDeployUrl, `${path}.rawDeployUrl`),',
-      '    finalUrl: expectUrl(record.finalUrl, `${path}.finalUrl`),',
-      '    deployedAtUtc: expectIsoUtc(record.deployedAtUtc, `${path}.deployedAtUtc`),',
-      '  }',
-      '}',
-      '',
-      'const validateCommit = (value, path) => {',
-      '  const record = expectObject(value, path)',
-      "  expectExactKeys(record, ['commitSha', 'modeLabel', 'deployedAtUtc', 'packages'], path)",
-      '  const packages = expectArray(record.packages, `${path}.packages`).map((pkg, index) => ({ value: pkg, path: `${path}.packages[${index}]` }))',
-      '  return {',
-      '    commitSha: expectString(record.commitSha, `${path}.commitSha`),',
-      '    modeLabel: expectString(record.modeLabel, `${path}.modeLabel`),',
-      '    deployedAtUtc: expectIsoUtc(record.deployedAtUtc, `${path}.deployedAtUtc`),',
-      '    packages,',
-      '  }',
-      '}',
-      '',
-      'const validateState = (value) => {',
-      "  const record = expectObject(value, 'state')",
-      "  expectExactKeys(record, ['_tag', 'schemaVersion', 'timeZone', 'packageOrder', 'commits'], 'state')",
-      '  if (record._tag !== stateTag) fail(`state._tag must be ${stateTag}`)',
-      '  if (![1, 2, schemaVersion].includes(record.schemaVersion)) fail(`state.schemaVersion must be 1, 2, or ${schemaVersion}`)',
-      '  if (record.timeZone !== timeZone) fail(`state.timeZone must be ${timeZone}`)',
-      '  const packageOrder = expectArray(record.packageOrder, "state.packageOrder").map((pkg, index) =>',
-      '    expectString(pkg, `state.packageOrder[${index}]`),',
-      '  )',
-      '  const commits = expectArray(record.commits, "state.commits").map((commit, index) => {',
-      '    const validatedCommit = validateCommit(commit, `state.commits[${index}]`)',
-      '    return {',
-      '      ...validatedCommit,',
-      '      packages: validatedCommit.packages.map((pkg, packageIndex) =>',
-      '        ((validatedPackage) => ({',
-      '          ...validatedPackage,',
-      '          deployedAtUtc: validatedPackage.deployedAtUtc ?? validatedCommit.deployedAtUtc,',
-      '        }))(',
-      '          record.schemaVersion === 1',
-      '            ? validatePackageV1(pkg.value, `state.commits[${index}].packages[${packageIndex}]`)',
-      '            : record.schemaVersion === 2',
-      '              ? validatePackageV2(pkg.value, `state.commits[${index}].packages[${packageIndex}]`)',
-      '              : validatePackageV3(pkg.value, `state.commits[${index}].packages[${packageIndex}]`),',
-      '        ),',
-      '      ),',
-      '    }',
-      '  })',
-      '  return { _tag: stateTag, schemaVersion, timeZone, packageOrder, commits }',
-      '}',
-      '',
-      'const formatBerlin = (isoUtc) => {',
-      "  const formatter = new Intl.DateTimeFormat('en-GB', {",
-      '    timeZone,',
-      "    year: 'numeric',",
-      "    month: '2-digit',",
-      "    day: '2-digit',",
-      "    hour: '2-digit',",
-      "    minute: '2-digit',",
-      '    hour12: false,',
-      "    timeZoneName: 'short',",
-      '  })',
-      '  const parts = Object.fromEntries(',
-      '    formatter.formatToParts(new Date(isoUtc)).filter((part) => part.type !== "literal").map((part) => [part.type, part.value]),',
-      '  )',
-      '  return `${parts.year}-${parts.month}-${parts.day} ${parts.hour}:${parts.minute} ${parts.timeZoneName}`',
-      '}',
-      '',
-      'const extractState = (body) => {',
-      '  const start = body.indexOf(statePrefix)',
-      '  if (start === -1) return undefined',
-      '  const end = body.indexOf(stateSuffix, start + statePrefix.length)',
-      '  if (end === -1) fail("existing managed comment is missing the state suffix marker")',
-      '  const raw = body.slice(start + statePrefix.length, end)',
-      '  return validateState(JSON.parse(raw))',
-      '}',
-      '',
-      'const renderHistoryUrl = (pkg) =>',
-      '  pkg.finalUrl === pkg.rawDeployUrl',
-      '    ? pkg.rawDeployUrl',
-      '    : `${pkg.rawDeployUrl}<br><sub>Alias: ${pkg.finalUrl}</sub>`',
-      '',
-      'const currentPackagesJson = expectString(process.env.STORYBOOK_PREVIEW_PACKAGES_JSON, "STORYBOOK_PREVIEW_PACKAGES_JSON")',
-      'const fallbackDeployedAtUtc = expectIsoUtc(process.env.STORYBOOK_PREVIEW_DEPLOYED_AT_UTC, "STORYBOOK_PREVIEW_DEPLOYED_AT_UTC")',
-      '',
-      'const currentPackages = expectArray(JSON.parse(currentPackagesJson), "currentPackages").map((pkg, index) =>',
-      '  validatePackageV3(pkg, `currentPackages[${index}]`),',
-      ')',
-      '',
-      'const deployedAtUtc = currentPackages.reduce(',
-      '  (latest, pkg) => (pkg.deployedAtUtc > latest ? pkg.deployedAtUtc : latest),',
-      '  fallbackDeployedAtUtc,',
-      ')',
-      '',
-      'currentPackages.forEach((pkg, index) => validatePackageV3(pkg, `currentPackages[${index}]`))',
-      '',
-      'if (currentPackages.length === 0) {',
-      '  writeFileSync(summaryPath, `## ${commentTitle}\\n\\n${noRowsMessage}\\n`)',
-      "  writeFileSync(commentBodyPath, '')",
-      "  writeFileSync(commentIdPath, '')",
-      '  process.exit(0)',
-      '}',
-      '',
-      "const comments = JSON.parse(readFileSync(commentsPath, 'utf8'))",
-      'if (!Array.isArray(comments)) fail("comments response must be an array")',
-      '',
-      'const existingComment = comments',
-      '  .filter(',
-      '    (comment) =>',
-      '      typeof comment?.body === "string" &&',
-      '      (comment.body.includes(managedMarker) || comment.body.startsWith(`## ${commentTitle}`)),',
-      '  )',
-      '  .at(-1)',
-      '',
-      'const existingState = existingComment ? extractState(existingComment.body) : undefined',
-      'const priorState = existingState ?? { _tag: stateTag, schemaVersion, timeZone, packageOrder: [], commits: [] }',
-      '',
-      'const commitSha = expectString(process.env.STORYBOOK_PREVIEW_COMMIT_SHA, "STORYBOOK_PREVIEW_COMMIT_SHA")',
-      'const modeLabel = expectString(process.env.STORYBOOK_PREVIEW_LABEL, "STORYBOOK_PREVIEW_LABEL")',
-      '',
-      'const nextPackageOrder = [...new Set([...currentPackages.map((pkg) => pkg.packageName), ...priorState.packageOrder])]',
-      'const nextState = {',
-      '  _tag: stateTag,',
-      '  schemaVersion,',
-      '  timeZone,',
-      '  packageOrder: nextPackageOrder,',
-      '  commits: [',
-      '    { commitSha, modeLabel, deployedAtUtc, packages: currentPackages },',
-      '    ...priorState.commits.filter((commit) => commit.commitSha !== commitSha),',
-      '  ].slice(0, maxCommits),',
-      '}',
-      '',
-      'const latestByPackage = new Map()',
-      'for (const commit of nextState.commits) {',
-      '  for (const pkg of commit.packages) {',
-      '    if (!latestByPackage.has(pkg.packageName)) {',
-      '      latestByPackage.set(pkg.packageName, pkg)',
-      '    }',
-      '  }',
-      '}',
-      '',
-      'const visibleLines = [',
-      '  `## ${commentTitle}`,',
-      "  '',",
-      '  `| Package | Latest URL | Last Deploy (Europe/Berlin) |`,',
-      '  `| --- | --- | --- |`,',
-      '  ...nextState.packageOrder.flatMap((packageName) => {',
-      '    const latest = latestByPackage.get(packageName)',
-      '    if (!latest) return []',
-      '    return [`| ${packageName} | ${latest.finalUrl} | ${formatBerlin(latest.deployedAtUtc)} |`]',
-      '  }),',
-      ']',
-      '',
-      'if (nextState.commits.length > 0) {',
-      "  visibleLines.push('', '<details>', '<summary>Per-Commit Deploy History</summary>', '')",
-      '  for (const commit of nextState.commits) {',
-      '    visibleLines.push(`### Commit \\`${commit.commitSha.slice(0, 7)}\\` · ${formatBerlin(commit.deployedAtUtc)}`)',
-      "    visibleLines.push('', '| Package | URL |', '| --- | --- |')",
-      '    for (const packageName of nextState.packageOrder) {',
-      '      const pkg = commit.packages.find((entry) => entry.packageName === packageName)',
-      '      if (!pkg) continue',
-      '      visibleLines.push(`| ${packageName} | ${renderHistoryUrl(pkg)} |`)',
-      '    }',
-      "    visibleLines.push('')",
-      '  }',
-      "  visibleLines.push('</details>')",
-      '}',
-      '',
-      'const body = [',
-      '  ...visibleLines,',
-      "  '',",
-      '  managedMarker,',
-      '  `${statePrefix}${JSON.stringify(nextState, null, 2)}${stateSuffix}`,',
-      "].join('\\n')",
-      '',
-      'writeFileSync(commentBodyPath, `${body}\\n`)',
-      'writeFileSync(summaryPath, `${visibleLines.join("\\n")}\\n`)',
-      'writeFileSync(commentIdPath, existingComment ? String(existingComment.id) : "")',
-    ].join('\n'),
-    'EOF',
-    'nix run nixpkgs#nodejs_24 -- /tmp/render-storybook-preview-comment.mjs "$comments_json" /tmp/storybook-preview-comment.md "$GITHUB_STEP_SUMMARY" /tmp/storybook-preview-comment-id.txt',
-    'comment_id="$(cat /tmp/storybook-preview-comment-id.txt)"',
-    'if [ ! -s /tmp/storybook-preview-comment.md ]; then',
-    '  exit 0',
-    'fi',
-    `export NIX_CONFIG="\${NIX_CONFIG:+$NIX_CONFIG$'\\n'}access-tokens = github.com=\${GH_TOKEN}"`,
-    'if [ "${{ github.event_name }}" != "pull_request" ]; then',
-    '  exit 0',
-    'fi',
-    'if [ -n "$comment_id" ]; then',
-    '  nix run nixpkgs#gh -- api "repos/$GH_REPO/issues/comments/$comment_id" --method PATCH --field body="$(cat /tmp/storybook-preview-comment.md)" >/dev/null',
-    'else',
-    '  nix run nixpkgs#gh -- api "repos/$GH_REPO/issues/${{ github.event.pull_request.number }}/comments" --method POST --field body="$(cat /tmp/storybook-preview-comment.md)" >/dev/null',
-    'fi',
-  ].join('\n'),
-})
+export const netlifyStorybookCommentStep = (site: string) =>
+  buildNetlifyStorybookCommentStep(site, deployModeScript)
