+++ genie/deploy-preview/netlify.ts
@@ -1,0 +1,366 @@
+type RunTasksBefore = (...tasks: [string, ...string[]]) => string
+
+export const netlifyDeployStep = (runDevenvTasksBefore: RunTasksBefore) => ({
+  id: 'deploy',
+  name: 'Deploy storybooks to Netlify',
+  shell: 'bash' as const,
+  run: [
+    'tmp_log="/tmp/netlify-storybook-deploy.log"',
+    'rm -f "$tmp_log"',
+    'if [ -z "${NETLIFY_AUTH_TOKEN:-}" ]; then',
+    '  echo "::notice::Skipping Netlify deploy (NETLIFY_AUTH_TOKEN not available)"',
+    '  exit 0',
+    'fi',
+    'if [ "${{ github.event_name }}" = "push" ] && [ "${{ github.ref }}" = "refs/heads/main" ]; then',
+    `  ${runDevenvTasksBefore('netlify:deploy', '--show-output', '--input', 'type=prod')} 2>&1 | tee "$tmp_log"`,
+    'elif [ "${{ github.event_name }}" = "pull_request" ]; then',
+    `  ${runDevenvTasksBefore('netlify:deploy', '--show-output', '--input', 'type=pr', '--input', 'pr=${{ github.event.pull_request.number }}')} 2>&1 | tee "$tmp_log"`,
+    'fi',
+    'deploy_exit=${PIPESTATUS[0]}',
+    'if [ "$deploy_exit" -ne 0 ]; then exit "$deploy_exit"; fi',
+    "cat > /tmp/extract-netlify-preview-packages.mjs <<'EOF'",
+    [
+      "import { readFileSync } from 'node:fs'",
+      '',
+      'const [logPath] = process.argv.slice(2)',
+      'const prefixes = ["DEPLOY_TASK_METADATA: ", "NETLIFY_DEPLOY_METADATA: "]',
+      'const packages = readFileSync(logPath, "utf8")',
+      '  .split("\\n")',
+      '  .flatMap((line) => {',
+      '    const prefix = prefixes.find((candidate) => line.includes(candidate))',
+      '    if (!prefix) return []',
+      '    const rawRecord = JSON.parse(line.slice(line.indexOf(prefix) + prefix.length))',
+      '    if (rawRecord.provider === "netlify") {',
+      '      return [{',
+      '        packageName: rawRecord.displayName ?? rawRecord.target,',
+      '        rawDeployUrl: rawRecord.rawDeployUrl,',
+      '        finalUrl: rawRecord.finalUrl,',
+      '        deployedAtUtc: rawRecord.deployedAtUtc,',
+      '      }]',
+      '    }',
+      '    if ("packageName" in rawRecord) return [rawRecord]',
+      '    return []',
+      '  })',
+      'const dedupedPackages = [...new Map(packages.map((pkg) => [pkg.packageName, pkg])).values()]',
+      'process.stdout.write(JSON.stringify(dedupedPackages))',
+    ].join('\n'),
+    'EOF',
+    'packages_json="$(node /tmp/extract-netlify-preview-packages.mjs "$tmp_log")"',
+    'printf \'packages_json<<__NETLIFY_PACKAGES__\\n%s\\n__NETLIFY_PACKAGES__\\n\' "$packages_json" >> "$GITHUB_OUTPUT"',
+  ].join('\n'),
+})
+
+export const netlifyStorybookCommentStep = (site: string, deployModeScript: string) => ({
+  name: 'Post storybook preview comment',
+  if: '${{ !cancelled() }}',
+  env: {
+    GH_TOKEN: '${{ github.token }}',
+    GH_REPO: '${{ github.repository }}',
+    STORYBOOK_PREVIEW_PACKAGES_JSON: '${{ steps.deploy.outputs.packages_json }}',
+  },
+  run: [
+    `site="${site}"`,
+    deployModeScript,
+    'if [ "$label" = "prod" ]; then suffix=""; else suffix="-pr-${{ github.event.pull_request.number }}"; fi',
+    'if [ "${{ github.event_name }}" = "pull_request" ]; then',
+    '  commit_sha="${{ github.event.pull_request.head.sha }}"',
+    'else',
+    '  commit_sha="${{ github.sha }}"',
+    'fi',
+    'export STORYBOOK_PREVIEW_LABEL="$label"',
+    'export STORYBOOK_PREVIEW_COMMIT_SHA="$commit_sha"',
+    'export STORYBOOK_PREVIEW_SITE="$site"',
+    'export STORYBOOK_PREVIEW_SUFFIX="$suffix"',
+    'export STORYBOOK_PREVIEW_DEPLOYED_AT_UTC="$(date -u +%Y-%m-%dT%H:%M:%SZ)"',
+    'comments_json="/tmp/storybook-preview-comments.json"',
+    'if [ "${{ github.event_name }}" = "pull_request" ]; then',
+    `  export NIX_CONFIG="\${NIX_CONFIG:+$NIX_CONFIG$'\\n'}access-tokens = github.com=\${GH_TOKEN}"`,
+    '  nix run nixpkgs#gh -- api "repos/$GH_REPO/issues/${{ github.event.pull_request.number }}/comments" --paginate > "$comments_json"',
+    'else',
+    '  printf \'[]\' > "$comments_json"',
+    'fi',
+    "cat > /tmp/render-storybook-preview-comment.mjs <<'EOF'",
+    [
+      "import { existsSync, readdirSync, readFileSync, writeFileSync } from 'node:fs'",
+      '',
+      'const commentTitle = "Storybook Previews"',
+      'const noRowsMessage = "No storybooks were deployed."',
+      'const managedMarker = "<!-- deploy-preview-comment:managed -->"',
+      'const statePrefix = "<!-- deploy-preview-comment:state\\n"',
+      'const stateSuffix = "\\n-->"',
+      "const stateTag = 'storybook-preview-comment-state'",
+      'const schemaVersion = 3',
+      "const timeZone = 'Europe/Berlin'",
+      'const maxCommits = 50',
+      '',
+      'const [commentsPath, commentBodyPath, summaryPath, commentIdPath] = process.argv.slice(2)',
+      '',
+      'const fail = (message) => {',
+      '  throw new Error(message)',
+      '}',
+      '',
+      'const expectObject = (value, path) => {',
+      "  if (typeof value !== 'object' || value === null || Array.isArray(value)) {",
+      '    fail(`${path} must be an object`)',
+      '  }',
+      '  return value',
+      '}',
+      '',
+      'const expectArray = (value, path) => {',
+      '  if (!Array.isArray(value)) {',
+      '    fail(`${path} must be an array`)',
+      '  }',
+      '  return value',
+      '}',
+      '',
+      'const expectString = (value, path) => {',
+      "  if (typeof value !== 'string' || value.length === 0) {",
+      '    fail(`${path} must be a non-empty string`)',
+      '  }',
+      '  return value',
+      '}',
+      '',
+      'const expectExactKeys = (record, keys, path) => {',
+      '  const actualKeys = Object.keys(record).sort()',
+      '  const expectedKeys = [...keys].sort()',
+      '  if (actualKeys.length !== expectedKeys.length || actualKeys.some((key, index) => key !== expectedKeys[index])) {',
+      '    fail(`${path} keys must be exactly: ${expectedKeys.join(", ")}`)',
+      '  }',
+      '}',
+      '',
+      'const expectUrl = (value, path) => {',
+      '  const string = expectString(value, path)',
+      "  if (!string.startsWith('https://')) {",
+      '    fail(`${path} must start with https://`)',
+      '  }',
+      '  return string',
+      '}',
+      '',
+      'const expectIsoUtc = (value, path) => {',
+      '  const string = expectString(value, path)',
+      '  if (!/^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$/.test(string) || Number.isNaN(Date.parse(string))) {',
+      '    fail(`${path} must be a valid UTC ISO timestamp`)',
+      '  }',
+      '  return string',
+      '}',
+      '',
+      'const validatePackageV1 = (value, path) => {',
+      '  const record = expectObject(value, path)',
+      "  expectExactKeys(record, ['packageName', 'url'], path)",
+      '  return {',
+      '    packageName: expectString(record.packageName, `${path}.packageName`),',
+      '    rawDeployUrl: expectUrl(record.url, `${path}.url`),',
+      '    finalUrl: expectUrl(record.url, `${path}.url`),',
+      '    deployedAtUtc: undefined,',
+      '  }',
+      '}',
+      '',
+      'const validatePackageV2 = (value, path) => {',
+      '  const record = expectObject(value, path)',
+      "  expectExactKeys(record, ['packageName', 'rawDeployUrl', 'finalUrl'], path)",
+      '  return {',
+      '    packageName: expectString(record.packageName, `${path}.packageName`),',
+      '    rawDeployUrl: expectUrl(record.rawDeployUrl, `${path}.rawDeployUrl`),',
+      '    finalUrl: expectUrl(record.finalUrl, `${path}.finalUrl`),',
+      '    deployedAtUtc: undefined,',
+      '  }',
+      '}',
+      '',
+      'const validatePackageV3 = (value, path) => {',
+      '  const record = expectObject(value, path)',
+      "  expectExactKeys(record, ['packageName', 'rawDeployUrl', 'finalUrl', 'deployedAtUtc'], path)",
+      '  return {',
+      '    packageName: expectString(record.packageName, `${path}.packageName`),',
+      '    rawDeployUrl: expectUrl(record.rawDeployUrl, `${path}.rawDeployUrl`),',
+      '    finalUrl: expectUrl(record.finalUrl, `${path}.finalUrl`),',
+      '    deployedAtUtc: expectIsoUtc(record.deployedAtUtc, `${path}.deployedAtUtc`),',
+      '  }',
+      '}',
+      '',
+      'const validateCommit = (value, path) => {',
+      '  const record = expectObject(value, path)',
+      "  expectExactKeys(record, ['commitSha', 'modeLabel', 'deployedAtUtc', 'packages'], path)",
+      '  const packages = expectArray(record.packages, `${path}.packages`).map((pkg, index) => ({ value: pkg, path: `${path}.packages[${index}]` }))',
+      '  return {',
+      '    commitSha: expectString(record.commitSha, `${path}.commitSha`),',
+      '    modeLabel: expectString(record.modeLabel, `${path}.modeLabel`),',
+      '    deployedAtUtc: expectIsoUtc(record.deployedAtUtc, `${path}.deployedAtUtc`),',
+      '    packages,',
+      '  }',
+      '}',
+      '',
+      'const validateState = (value) => {',
+      "  const record = expectObject(value, 'state')",
+      "  expectExactKeys(record, ['_tag', 'schemaVersion', 'timeZone', 'packageOrder', 'commits'], 'state')",
+      '  if (record._tag !== stateTag) fail(`state._tag must be ${stateTag}`)',
+      '  if (![1, 2, schemaVersion].includes(record.schemaVersion)) fail(`state.schemaVersion must be 1, 2, or ${schemaVersion}`)',
+      '  if (record.timeZone !== timeZone) fail(`state.timeZone must be ${timeZone}`)',
+      '  const packageOrder = expectArray(record.packageOrder, "state.packageOrder").map((pkg, index) =>',
+      '    expectString(pkg, `state.packageOrder[${index}]`),',
+      '  )',
+      '  const commits = expectArray(record.commits, "state.commits").map((commit, index) => {',
+      '    const validatedCommit = validateCommit(commit, `state.commits[${index}]`)',
+      '    return {',
+      '      ...validatedCommit,',
+      '      packages: validatedCommit.packages.map((pkg, packageIndex) =>',
+      '        ((validatedPackage) => ({',
+      '          ...validatedPackage,',
+      '          deployedAtUtc: validatedPackage.deployedAtUtc ?? validatedCommit.deployedAtUtc,',
+      '        }))(',
+      '          record.schemaVersion === 1',
+      '            ? validatePackageV1(pkg.value, `state.commits[${index}].packages[${packageIndex}]`)',
+      '            : record.schemaVersion === 2',
+      '              ? validatePackageV2(pkg.value, `state.commits[${index}].packages[${packageIndex}]`)',
+      '              : validatePackageV3(pkg.value, `state.commits[${index}].packages[${packageIndex}]`),',
+      '        ),',
+      '      ),',
+      '    }',
+      '  })',
+      '  return { _tag: stateTag, schemaVersion, timeZone, packageOrder, commits }',
+      '}',
+      '',
+      'const formatBerlin = (isoUtc) => {',
+      "  const formatter = new Intl.DateTimeFormat('en-GB', {",
+      '    timeZone,',
+      "    year: 'numeric',",
+      "    month: '2-digit',",
+      "    day: '2-digit',",
+      "    hour: '2-digit',",
+      "    minute: '2-digit',",
+      '    hour12: false,',
+      "    timeZoneName: 'short',",
+      '  })',
+      '  const parts = Object.fromEntries(',
+      '    formatter.formatToParts(new Date(isoUtc)).filter((part) => part.type !== "literal").map((part) => [part.type, part.value]),',
+      '  )',
+      '  return `${parts.year}-${parts.month}-${parts.day} ${parts.hour}:${parts.minute} ${parts.timeZoneName}`',
+      '}',
+      '',
+      'const extractState = (body) => {',
+      '  const start = body.indexOf(statePrefix)',
+      '  if (start === -1) return undefined',
+      '  const end = body.indexOf(stateSuffix, start + statePrefix.length)',
+      '  if (end === -1) fail("existing managed comment is missing the state suffix marker")',
+      '  const raw = body.slice(start + statePrefix.length, end)',
+      '  return validateState(JSON.parse(raw))',
+      '}',
+      '',
+      'const renderHistoryUrl = (pkg) =>',
+      '  pkg.finalUrl === pkg.rawDeployUrl',
+      '    ? pkg.rawDeployUrl',
+      '    : `${pkg.rawDeployUrl}<br><sub>Alias: ${pkg.finalUrl}</sub>`',
+      '',
+      'const currentPackagesJson = expectString(process.env.STORYBOOK_PREVIEW_PACKAGES_JSON, "STORYBOOK_PREVIEW_PACKAGES_JSON")',
+      'const fallbackDeployedAtUtc = expectIsoUtc(process.env.STORYBOOK_PREVIEW_DEPLOYED_AT_UTC, "STORYBOOK_PREVIEW_DEPLOYED_AT_UTC")',
+      '',
+      'const currentPackages = expectArray(JSON.parse(currentPackagesJson), "currentPackages").map((pkg, index) =>',
+      '  validatePackageV3(pkg, `currentPackages[${index}]`),',
+      ')',
+      '',
+      'const deployedAtUtc = currentPackages.reduce(',
+      '  (latest, pkg) => (pkg.deployedAtUtc > latest ? pkg.deployedAtUtc : latest),',
+      '  fallbackDeployedAtUtc,',
+      ')',
+      '',
+      'currentPackages.forEach((pkg, index) => validatePackageV3(pkg, `currentPackages[${index}]`))',
+      '',
+      'if (currentPackages.length === 0) {',
+      '  writeFileSync(summaryPath, `## ${commentTitle}\\n\\n${noRowsMessage}\\n`)',
+      "  writeFileSync(commentBodyPath, '')",
+      "  writeFileSync(commentIdPath, '')",
+      '  process.exit(0)',
+      '}',
+      '',
+      "const comments = JSON.parse(readFileSync(commentsPath, 'utf8'))",
+      'if (!Array.isArray(comments)) fail("comments response must be an array")',
+      '',
+      'const existingComment = comments',
+      '  .filter(',
+      '    (comment) =>',
+      '      typeof comment?.body === "string" &&',
+      '      (comment.body.includes(managedMarker) || comment.body.startsWith(`## ${commentTitle}`)),',
+      '  )',
+      '  .at(-1)',
+      '',
+      'const existingState = existingComment ? extractState(existingComment.body) : undefined',
+      'const priorState = existingState ?? { _tag: stateTag, schemaVersion, timeZone, packageOrder: [], commits: [] }',
+      '',
+      'const commitSha = expectString(process.env.STORYBOOK_PREVIEW_COMMIT_SHA, "STORYBOOK_PREVIEW_COMMIT_SHA")',
+      'const modeLabel = expectString(process.env.STORYBOOK_PREVIEW_LABEL, "STORYBOOK_PREVIEW_LABEL")',
+      '',
+      'const nextPackageOrder = [...new Set([...currentPackages.map((pkg) => pkg.packageName), ...priorState.packageOrder])]',
+      'const nextState = {',
+      '  _tag: stateTag,',
+      '  schemaVersion,',
+      '  timeZone,',
+      '  packageOrder: nextPackageOrder,',
+      '  commits: [',
+      '    { commitSha, modeLabel, deployedAtUtc, packages: currentPackages },',
+      '    ...priorState.commits.filter((commit) => commit.commitSha !== commitSha),',
+      '  ].slice(0, maxCommits),',
+      '}',
+      '',
+      'const latestByPackage = new Map()',
+      'for (const commit of nextState.commits) {',
+      '  for (const pkg of commit.packages) {',
+      '    if (!latestByPackage.has(pkg.packageName)) {',
+      '      latestByPackage.set(pkg.packageName, pkg)',
+      '    }',
+      '  }',
+      '}',
+      '',
+      'const visibleLines = [',
+      '  `## ${commentTitle}`,',
+      "  '',",
+      '  `| Package | Latest URL | Last Deploy (Europe/Berlin) |`,',
+      '  `| --- | --- | --- |`,',
+      '  ...nextState.packageOrder.flatMap((packageName) => {',
+      '    const latest = latestByPackage.get(packageName)',
+      '    if (!latest) return []',
+      '    return [`| ${packageName} | ${latest.finalUrl} | ${formatBerlin(latest.deployedAtUtc)} |`]',
+      '  }),',
+      ']',
+      '',
+      'if (nextState.commits.length > 0) {',
+      "  visibleLines.push('', '<details>', '<summary>Per-Commit Deploy History</summary>', '')",
+      '  for (const commit of nextState.commits) {',
+      '    visibleLines.push(`### Commit \\`${commit.commitSha.slice(0, 7)}\\` · ${formatBerlin(commit.deployedAtUtc)}`)',
+      "    visibleLines.push('', '| Package | URL |', '| --- | --- |')",
+      '    for (const packageName of nextState.packageOrder) {',
+      '      const pkg = commit.packages.find((entry) => entry.packageName === packageName)',
+      '      if (!pkg) continue',
+      '      visibleLines.push(`| ${packageName} | ${renderHistoryUrl(pkg)} |`)',
+      '    }',
+      "    visibleLines.push('')",
+      '  }',
+      "  visibleLines.push('</details>')",
+      '}',
+      '',
+      'const body = [',
+      '  ...visibleLines,',
+      "  '',",
+      '  managedMarker,',
+      '  `${statePrefix}${JSON.stringify(nextState, null, 2)}${stateSuffix}`,',
+      "].join('\\n')",
+      '',
+      'writeFileSync(commentBodyPath, `${body}\\n`)',
+      'writeFileSync(summaryPath, `${visibleLines.join("\\n")}\\n`)',
+      'writeFileSync(commentIdPath, existingComment ? String(existingComment.id) : "")',
+    ].join('\n'),
+    'EOF',
+    'nix run nixpkgs#nodejs_24 -- /tmp/render-storybook-preview-comment.mjs "$comments_json" /tmp/storybook-preview-comment.md "$GITHUB_STEP_SUMMARY" /tmp/storybook-preview-comment-id.txt',
+    'comment_id="$(cat /tmp/storybook-preview-comment-id.txt)"',
+    'if [ ! -s /tmp/storybook-preview-comment.md ]; then',
+    '  exit 0',
+    'fi',
+    `export NIX_CONFIG="\${NIX_CONFIG:+$NIX_CONFIG$'\\n'}access-tokens = github.com=\${GH_TOKEN}"`,
+    'if [ "${{ github.event_name }}" != "pull_request" ]; then',
+    '  exit 0',
+    'fi',
+    'if [ -n "$comment_id" ]; then',
+    '  nix run nixpkgs#gh -- api "repos/$GH_REPO/issues/comments/$comment_id" --method PATCH --field body="$(cat /tmp/storybook-preview-comment.md)" >/dev/null',
+    'else',
+    '  nix run nixpkgs#gh -- api "repos/$GH_REPO/issues/${{ github.event.pull_request.number }}/comments" --method POST --field body="$(cat /tmp/storybook-preview-comment.md)" >/dev/null',
+    'fi',
+  ].join('\n'),
+})
